pulling remote specific questions out to remote specific runners
definitely needs to be refactored at some point
