+++ src/Nagger.Interfaces/IRemoteRunner.cs
@@ -1,0 +1,9 @@
+﻿namespace Nagger.Interfaces
+{
+    using Models;
+
+    public interface IRemoteRunner
+    {
+        Task AskForTask();
+    }
+}
