+++ src/Nagger.Services/Nagger.Services.csproj
@@ -41,6 +41,7 @@
     <Compile Include="ExtensionMethods\StringExtensions.cs" />
     <Compile Include="ExtensionMethods\TimeSpanExtensions.cs" />
     <Compile Include="ConsoleInputService.cs" />
+    <Compile Include="JIRA\JiraRunner.cs" />
     <Compile Include="NaggerRunner.cs" />
     <Compile Include="ProjectService.cs" />
     <Compile Include="SettingsService.cs" />
